Store on-page position in bookmarks created from "Add bookmark" menu item. Store position in navigtion history for "GoTo Page" dialog.
git-svn-id: http://ebookdroid.googlecode.com/svn/trunk@1204 714301b5-5bf2-3aa6-4ccc-ecebab0412e3

+++ src/org/ebookdroid/common/settings/books/Bookmark.java
@@ -10,7 +10,7 @@ public class Bookmark {
     public float offsetX;
     public float offsetY;
 
-    public Bookmark(final String name, final PageIndex page, int offsetX, int offsetY) {
+    public Bookmark(final String name, final PageIndex page, float offsetX, float offsetY) {
         this(false, name, page, offsetX, offsetY);
     }
